improvement on apis

+++ flask_todo/resources/task.py
@@ -11,23 +11,21 @@ INCOMING_DATE_FMT = '%d/%m/%Y %H:%M:%S'
 class TaskList(Resource):
     @jwt_required()
     def get(self):
+        """List all of the tasks for one user."""
         username = current_identity
-        return {'Tasks': username.tasks}
+        return {'Tasks': [task.to_dict()for task in username.tasks]}
 
 
 class CreateTask(Resource):
     @jwt_required()
     def post(self):
         username = current_identity
-        print(username)
-
         parser = reqparse.RequestParser()
         parser.add_argument('name', required=True, help="This field cannot be blank!")
         parser.add_argument('note', required=True, help="This field cannot be blank!")
         parser.add_argument('due-date', help="This field cannot be blank!")
         parser.add_argument('completed', help="This field cannot be blank!")
         data = parser.parse_args()
-
         try:
             task = TaskModel(name=data['name'], note=data['note'],user=username) 
             task.save_to_db()
@@ -35,22 +33,27 @@ class CreateTask(Resource):
         except:
             return {'message': 'Something wrong happend.'}, 400
 
+
 class Task(Resource):
+    @jwt_required()
     def get(self, task_id):
-        """Get the detail for an individual profile."""
-        task = TaskModel.find_by_id(task_id)
-        if task:
+        """Get the detail for one task if that task belongs to the provided user."""
+        task = TaskModel.query.filter_by(id=task_id).first()
+        current_user_tasks = current_identity.tasks
+        if task and task in current_user_tasks:
             return task.to_dict()
         return {'message': 'Task not found'}, 404
 
     def put(self, task_id):
+        """Update one task if that task belongs to the provided user."""
         pass
 
+    @jwt_required()
     def delete(self, task_id):
-        """."""
+        """Delete one task if that task belongs to the provided user."""
         task = TaskModel.query.filter_by(id=task_id).first()
-        if task:
+        current_user_tasks = current_identity.tasks
+        if task and task in current_user_tasks:
             task.delete_from_db()
             return {'message': f'Task {task_id} has been deleted.'}, 201
         return {'message': 'Task not found'}, 404
-
